files/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf
psycopg2-binary